Fix sparse vector nnz to use size_t, not uint32
Per zvec maintainer (issue #443): the sparse vector format is
[nnz:size_t][indices:uint32*nnz][values:float*nnz]. We were writing
nnz as uint32 (4 bytes) but it should be size_t (8 bytes on x64).

Also noted: the reranker API will change in the next zvec release.
Standalone zvec_reranker_create_rrf/create_weighted are being replaced
with zvec_multi_query_set_rerank_rrf/set_rerank_weighted. Our current
code works with the main-branch DLL but will need updating when the
official release ships.

+++ src/Zvec/ZvecDocument.cs
@@ -118,15 +118,16 @@ public class ZvecDocument : IDisposable
         ObjectDisposedException.ThrowIf(_disposed, this);
 
         int nnz = sparseVector.Count;
-        // Format: [nnz:uint32][indices:uint32 * nnz][values:float * nnz]
-        int bufferSize = sizeof(uint) + nnz * sizeof(uint) + nnz * sizeof(float);
+        // Format per zvec maintainer: [nnz:size_t][indices:uint32 * nnz][values:float * nnz]
+        // size_t is 8 bytes on 64-bit platforms
+        int bufferSize = sizeof(nuint) + nnz * sizeof(uint) + nnz * sizeof(float);
         var buffer = new byte[bufferSize];
 
         fixed (byte* ptr = buffer)
         {
-            *(uint*)ptr = (uint)nnz;
-            uint* indices = (uint*)(ptr + sizeof(uint));
-            float* values = (float*)(ptr + sizeof(uint) + nnz * sizeof(uint));
+            *(nuint*)ptr = (nuint)nnz;
+            uint* indices = (uint*)(ptr + sizeof(nuint));
+            float* values = (float*)(ptr + sizeof(nuint) + nnz * sizeof(uint));
 
             int i = 0;
             foreach (var kv in sparseVector)
